[refactor] split JobConfig and ConfigManager into two files (#1442)
This PR creates a new folder `torchtitan/config` to host `job_config.py`
and `manager.py`, for the reasons below:
- Both are complicated enough to worth their own files.
- The convention in torchtitan to extend custom `JobConfig` is to create
a file under model folder called `job_config.py` (see
https://github.com/pytorch/torchtitan/blob/main/torchtitan/experiments/flux/job_config.py).
This PR makes the origin `JobConfig` consistent with that convention.
- (minor) Creating a more succinct `torchtitan.config` namespace is more
readable than importing from `torchtitan.config_manager`.
