+++ torchtitan/components/ft.py
@@ -13,7 +13,7 @@ import torch
 import torch.distributed as dist
 from torch.distributed._composable.fsdp.fully_shard import FSDPModule
 from torch.distributed.distributed_c10d import ReduceOp
-from torchtitan.config_manager import FaultTolerance as FTConfig
+from torchtitan.config.job_config import FaultTolerance as FTConfig
 
 if importlib.util.find_spec("torchft") is not None:
     import torchft as ft
