+++ torchtitan/experiments/flux/tests/unit_tests/test_flux_dataloader.py
@@ -10,7 +10,7 @@ import torch
 
 from datasets import load_dataset
 
-from torchtitan.config_manager import ConfigManager
+from torchtitan.config import ConfigManager
 from torchtitan.experiments.flux.dataset.flux_dataset import (
     _cc12m_wds_data_processor,
     build_flux_dataloader,
