+++ torchtitan/experiments/llama4/model/args.py
@@ -9,7 +9,7 @@ from dataclasses import dataclass
 
 from torch import nn
 
-from torchtitan.config_manager import JobConfig
+from torchtitan.config import JobConfig
 from torchtitan.protocols.train_spec import BaseModelArgs
 from torchtitan.tools.logging import logger
 from torchtitan.tools.utils import has_cuda_capability
