+++ torchtitan/models/deepseek_v3/infra/pipeline.py
@@ -20,7 +20,7 @@ from torch.distributed.pipelining.schedules import (
 )
 
 from torchtitan.components.loss import LossFunction
-from torchtitan.config_manager import JobConfig
+from torchtitan.config import JobConfig
 from torchtitan.distributed import ParallelDims
 from torchtitan.distributed.pipeline import build_pipeline_schedule, stage_ids_this_rank
 from torchtitan.protocols.train_spec import ParallelizeFunction
