+++ torchtitan/models/deepseek_v3/model/args.py
@@ -12,7 +12,7 @@ from typing import Literal
 
 from torch import nn
 
-from torchtitan.config_manager import JobConfig
+from torchtitan.config import JobConfig
 from torchtitan.protocols.train_spec import BaseModelArgs
 from torchtitan.tools.logging import logger
 from torchtitan.tools.utils import has_cuda_capability
