+++ torchtitan/tools/profiling.py
@@ -11,7 +11,7 @@ import time
 
 import torch
 
-from torchtitan.config_manager import JobConfig
+from torchtitan.config import JobConfig
 from torchtitan.tools.logging import logger
 
 # the number of warmup steps before the active step in each profiling cycle
